requirements.txt
ipykernel

jupyter
notebook

numpy
pandas
scikit-learn

langchain
langchain-core
langchain_community

langchain-openai
openai

langchain-huggingface
transformers
huggingface-hub

langgraph

python-dotenv

streamlit

ddgs